backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "appwrite>=14.1.0",
    "fastapi>=0.128.0",
    "openai>=2.16.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "twilio>=9.10.0",
    "uvicorn>=0.40.0",
]
